30.06.2022

+++ ticketSite/ticketApp/views.py
@@ -1,19 +1,30 @@
 from django.contrib.auth import login, authenticate, logout
 from django.contrib.auth.forms import AuthenticationForm
-from django.contrib import messages
 from django.shortcuts import render, redirect
-from .forms import NewAccountForm, CreateTicketForm
+from .forms import *
 from .models import *
 
 
 def create_ticket(request):
 	if request.user.is_authenticated:
 		if request.method == "POST":
-			form = CreateTicketForm(request.POST)
-			if form.is_valid():
-				form.save()
-				return redirect('index')
-		form = CreateTicketForm
+			if request.user.is_superuser:
+				form = CreateTicketFormAdmin(request.POST)
+				if form.is_valid():
+					form.save()
+					return redirect('index')
+			else:
+				form = CreateTicketForm(request.POST)
+				if form.is_valid():
+					ticket = form.save(commit=False)
+					ticket.assignee = request.user
+					ticket.save()
+					return redirect('index')
+		else:
+			if request.user.is_superuser:
+				form = CreateTicketFormAdmin(request.POST)
+			else:
+				form = CreateTicketForm
 		return render(request, 'CreateTicket.html', {'form': form})
 	return render(request, 'index.html')
 
@@ -21,19 +32,30 @@ def create_ticket(request):
 def edit_ticket(request, ticket_id):
 	if request.user.is_authenticated:
 		ticket = Ticket.objects.get(pk=ticket_id)
-		form = CreateTicketForm(request.POST or None, instance=ticket)
-		if form.is_valid():
-			form.save()
+		if request.user.is_superuser:
+			form = CreateTicketFormAdmin(request.POST)
+			if form.is_valid():
+				form.save()
+				return redirect('index')
+		else:
+			if request.user == ticket.assignee:
+				form = CreateTicketForm(request.POST or None, instance=ticket)
+				if form.is_valid():
+					form.save()
+					return redirect('index')
+				return render(request, 'EditTicket.html', {'ticket': ticket, 'form': form})
 			return redirect('index')
-		return render(request, 'EditTicket.html', {'ticket': ticket, 'form': form})
 	return render(request, 'index.html')
 
 
 def delete_ticket(request, ticket_id):
 	if request.user.is_authenticated:
 		ticket = Ticket.objects.get(pk=ticket_id)
-		ticket.delete()
-		return redirect('index')
+		if request.user == ticket.assignee or request.user.is_superuser:
+			ticket.delete()
+			return redirect('index')
+		else:
+			return redirect('index')
 	return render(request, 'index.html')
 
 
@@ -45,7 +67,10 @@ def index(request):
 def ticket_by_id(request, ticket_id):
 	if request.user.is_authenticated:
 		ticket = Ticket.objects.get(pk=ticket_id)
-		return render(request, 'ticket_by_id.html', {'ticket': ticket})
+		if request.user == ticket.assignee or request.user.is_superuser:
+			return render(request, 'ticket_by_id.html', {'ticket': ticket})
+		else:
+			return redirect('index')
 	return render(request, 'index.html')
 
 
@@ -79,7 +104,6 @@ def login_request(request):
 				login(request, user)
 				return redirect('index')
 			else:
-				messages.success(request, ("There Was An Error Logging In, Try Again..."))
 				return redirect('login_request')
 	form = AuthenticationForm()
 	return render(request, "login_request.html", {'form': form})
